Flatten components list and pull out InteractionsPanel from Panel
