+++ addons/interactions/src/components/StatusIcon.stories.tsx
